refactor(common): add DOM/clipboard utilities and standardize CSS classes
- Add new DOM utilities to domUtils.js: clearElement, showElement,
  hideElement, setDisplayVisible, setVisibilityState, setDatasetBatch
- Create clipboardUtils.js with copyTextToClipboard and copyWithFeedback
- Re-export clipboard utilities from progressView/utils.js for compatibility
- Standardize status-indicator CSS classes to use is-* prefix pattern
  (is-ready, is-running, is-error, is-stopped, is-waiting)
